rewrite linearly_independent
linearly_independent used to generate the whole set generated by the given words, and check if it had as many elements as expected. That can be quite slow.

Instead, notice that if the set is not linearly independent, then there's a non-trivial solution to
	l_1w_1 + l_2w_2 + l_3w_3 + ... +l_nw_n = 0

So, iterate through all choices of the l_i, and if you ever get zero, the words aren't linearly independent.